add ed25519 implementation
This adds the ed25519 implementation. It implements most of the API,
apart from three functions.

Both MarshalPem methods write the raw bytes to the channel instead of
exporting it into ASN.1. Maybe this needs to be fixed.
The PrivateKey() function does not return a crypto.PrivateKey at the
moment, as the ed25519 private key does not implement that interface.